clean up flow types for modals
move internal flow types into their respective internal modules
create a flow type for ModalFocusTrap
